+++ src/MyServiceBus.AmazonSqs/AmazonSqsReceiveEndpointConfigurator.cs
@@ -71,7 +71,8 @@ public sealed class AmazonSqsReceiveEndpointConfigurator
             binding.EntityName = _entityNameResolver(binding.MessageType);
 
         consumer.PrefetchCount = _prefetchCount is null ? null : checked((ushort)_prefetchCount.Value);
-        consumer.ConcurrentMessageLimit = _concurrentMessageLimit;
+        if (_concurrentMessageLimit is not null)
+            consumer.ConcurrentMessageLimit = _concurrentMessageLimit;
         consumer.SerializerType = _serializerType;
         if (_retryCount.HasValue)
         {