+++ src/MyServiceBus.AzureServiceBus/AzureServiceBusReceiveEndpointConfigurator.cs
@@ -84,7 +84,8 @@ public sealed class AzureServiceBusReceiveEndpointConfigurator
         }
 
         consumer.PrefetchCount = _prefetchCount is null ? null : checked((ushort)_prefetchCount.Value);
-        consumer.ConcurrentMessageLimit = _concurrentMessageLimit;
+        if (_concurrentMessageLimit is not null)
+            consumer.ConcurrentMessageLimit = _concurrentMessageLimit;
         consumer.SerializerType = _serializerType;
 
         if (_retryCount.HasValue)