+++ src/MyServiceBus/IConsumerConfigurator.cs
@@ -2,4 +2,7 @@ namespace MyServiceBus;
 
 public interface IConsumerConfigurator<T> where T : class, IConsumer
 {
+    string? EndpointName { get; set; }
+
+    int? ConcurrentMessageLimit { get; set; }
 }