+++ src/MyServiceBus/IConsumerRegistrationConfigurator.cs
@@ -2,4 +2,5 @@ namespace MyServiceBus;
 
 public interface IConsumerRegistrationConfigurator<T> where T : class, IConsumer
 {
+    Topology.ConsumerDefinitionModel Definition { get; }
 }